misc/HtmlTestSuite/results/test-0450.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSSStyleSheet baseURL</title>
<link href="mailto:enordin@mozilla.com" rel="author" title="Erik Nordin"/>
<link href="https://github.com/WICG/construct-stylesheets/issues/95#issuecomment-593545252" rel="help"/>
</head><body><div id="target"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script>

function currentLocation() {
  const sections = location.href.split("/")
  sections.pop();
  return sections.join("/");
}

test(() =&gt; {
  const span = document.createElement("span");
  target.appendChild(span);
  span.attachShadow({ mode: "open" })
  const shadowDiv = document.createElement("div");
  span.shadowRoot.appendChild(shadowDiv);

  const fileName = "example.png"
  const baseURL = `${location.origin}/custom/path/`;
  const fullURL = `${baseURL}${fileName}`;

  const sheet = new CSSStyleSheet({ baseURL });
  span.shadowRoot.adoptedStyleSheets = [sheet];

  sheet.replaceSync(`* { background-image: url("${fileName}"); }`);
  const styleFromRelative = getComputedStyle(shadowDiv).backgroundImage;

  sheet.replaceSync(`* { background-image: url("${fullURL}"); }`);
  const styleFromFull = getComputedStyle(shadowDiv).backgroundImage;

  assert_equals(styleFromRelative, styleFromFull);
}, "Constructing sheet with custom base URL ueses that URL for CSS rules");

test(() =&gt; {
  const span = document.createElement("span");
  target.appendChild(span);
  span.attachShadow({ mode: "open" })
  const shadowDiv = document.createElement("div");
  span.shadowRoot.appendChild(shadowDiv);

  const fileName = "example.png"
  const baseURL = "custom/path/";
  const fullURL = `${currentLocation()}/${baseURL}${fileName}`;

  const sheet = new CSSStyleSheet({ baseURL });
  span.shadowRoot.adoptedStyleSheets = [sheet];

  sheet.replaceSync(`* { background-image: url("${fileName}"); }`);
  const styleFromRelative = getComputedStyle(shadowDiv).backgroundImage;

  sheet.replaceSync(`* { background-image: url("${fullURL}"); }`);
  const styleFromFull = getComputedStyle(shadowDiv).backgroundImage;

  assert_equals(styleFromRelative, styleFromFull);
}, "Constructing sheet with relative URL adds to the constructor document's base URL");

test(() =&gt; {
  assert_throws_dom("NotAllowedError", () =&gt; { new CSSStyleSheet({ baseURL: "https://test:test/"}) });
}, "Constructing sheet with invalid base URL throws a NotAllowedError");

</script>
</body></html>
